requirements.txt
Flask
python-dotenv
pycryptodome
